+++ app/src/main/res/layout/adapter_rv.xml
@@ -1,56 +1,61 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:padding="8dp"
     android:clipToPadding="false"
-    >
-        <android.support.v7.widget.RecyclerView
+    android:orientation="vertical"
+    android:padding="8dp"
+    android:paddingEnd="0dp">
+
+    <android.support.v7.widget.RecyclerView
+        android:id="@+id/rv"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" />
+
+    <android.support.v7.widget.CardView xmlns:card_view="http://schemas.android.com/apk/res-auto"
+        android:id="@+id/card1"
+        android:layout_width="match_parent"
+        android:layout_height="200dp"
+        android:layout_marginEnd="24dp"
+        card_view:cardCornerRadius="0dp"
+        card_view:cardElevation="4dp">
+
+        <RelativeLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:id="@+id/rv"
-            />
-        <android.support.v7.widget.CardView
-            xmlns:card_view="http://schemas.android.com/apk/res-auto"
-            android:id="@+id/card1"
-            android:layout_width="300dp"
-            android:layout_height="200dp"
-            card_view:cardElevation="6dp"
-            card_view:cardCornerRadius="0dp">
-            <RelativeLayout
-                android:layout_width="300dp"
-                android:layout_height="200dp"
-                android:padding="16dp">
-                <TextView
-                    android:id="@+id/card1name"
-                    android:text="Liana Alvarez"
-                    android:textSize="20dp"
-                    android:fontFamily="sans-serif-medium"
-                    android:textColor="#000000"
-                    android:layout_width="200dp"
-                    android:layout_height="40dp"
-                    />
-                <TextView
-                    android:layout_alignTop="@+id/card1name"
-                    android:layout_alignBaseline="@+id/card1name"
-                    android:id="@+id/card1time"
-                    android:text="3:00pm"
-                    android:textColor="#D5D5D5"
-                    android:textSize="16dp"
-                    android:layout_toRightOf="@id/card1name"
-                    android:layout_width="100dp"
-                    android:layout_height="40dp" />
-                <TextView
-                    android:id="@+id/card1review"
-                    android:text="This place sucks."
-                    android:textColor="#D5D5D5"
-                    android:textSize="16dp"
-                    android:layout_below="@+id/card1name"
-                    android:layout_width="300dp"
-                    android:layout_height="200dp"
-                    android:layout_alignEnd="@+id/card1time"
-                    android:layout_alignParentStart="true" />
-            </RelativeLayout>
-        </android.support.v7.widget.CardView>
+            android:padding="16dp">
+
+            <TextView
+                android:id="@+id/card1name"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_alignParentStart="true"
+                android:layout_toLeftOf="@+id/card1time"
+                android:fontFamily="sans-serif-medium"
+                android:text="Liana Alvarez"
+                android:textColor="#000000"
+                android:textSize="18dp" />
+
+            <TextView
+                android:id="@+id/card1time"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_alignBaseline="@+id/card1name"
+                android:layout_alignParentEnd="true"
+                android:layout_marginStart="12dp"
+                android:text="3:00pm"
+                android:textColor="#9B9B9B"
+                android:textSize="18dp" />
+
+            <TextView
+                android:id="@+id/card1review"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_below="@+id/card1name"
+                android:layout_marginTop="8dp"
+                android:text="This place sucks."
+                android:textColor="#4A4A4A"
+                android:textSize="18dp" />
+        </RelativeLayout>
+    </android.support.v7.widget.CardView>
 </RelativeLayout>
